Added ACL for the "ldap-client" role; this completes this role's config.
Added 2 new tasks: add a line to /etc/pam.d/common-account to
require access control, and set up ACLs in /etc/security/access.conf
using a template file.

The ldap groups are added using a group variable defined in
group_vars/drbu and dvgs. The "drbu-guest" group is taken care of
by checking whether the host is in the drbu-cs-lab-guest group.
Later if needed, we can rename the group drbu-guest-machines if
they are not limited to located in the cs lab.

Tested on asus-test.