Fixed timed_internet errors with better understanding on how it works.
dvbs' IP range encloses dvbs_cc_and_student_laptops', so it did not
work. Made them non-overlapping. Works cleanly now but a bit hard
to maintain.

dvgs and dvgs_cslab had a similar issue in IP range definition, but
amazingly their time limits are so defined that the problem is not
exposed. :-) But we better keep this in mind.